refactor(backend): port definition repo and drop PortMapping ORM
- Remove PortMapping model and GraphRunner/ComponentInstance relationships
- Rename port_mapping_repository to port_definition_repository (git mv); keep PortDefinition helpers only
- Point update_graph_service and output_port_instance_sync at the new module
- Delete port_mapping_schema.py
